Criação do crud TEMA

+++ src/postagem/services/postagem.service.ts
@@ -4,6 +4,7 @@ import { Postagem } from "../entities/postagem.entity";
 import { InjectRepository } from "@nestjs/typeorm";
 import { waitForDebugger } from "inspector";
 import { DeleteResult } from "typeorm/browser";
+import { TemaService } from "../../tema/services/tema.service";
 
 @Injectable() // Indica que a classe é de serviço e pode ser injetada em outras classes
 export class PostagemService{
@@ -11,17 +12,27 @@ export class PostagemService{
     // iniciando ferramentas para a classe de serviço
     constructor(
         @InjectRepository(Postagem) // Pode chamar os metodos de uma classe Repository
-        private postagemRepository: Repository<Postagem>
+        private postagemRepository: Repository<Postagem>,
+        private temaService: TemaService
     ){}
 
     // Procura todas as postagens
     async findAll(): Promise<Postagem[]>{
-        return await this.postagemRepository.find()
+        return await this.postagemRepository.find({
+            relations: {
+                tema: true
+            }
+        })
     }
 
     async findById(id: number): Promise<Postagem>{
         const postagem = await this.postagemRepository.findOne({
-            where: { id }
+            where: { 
+                id 
+            },
+            relations: {
+                tema: true
+            }
         })
         
         if(!postagem) {
@@ -35,24 +46,58 @@ export class PostagemService{
         return await this.postagemRepository.find({
             where: {
                 titulo: ILike(`%${titulo}%`)
+            },
+            relations: {
+                tema: true
             }
         })
     }
 
     async createPost(postagem: Postagem): Promise<Postagem>{
+
+        if(postagem.tema){
+
+            let tema = await this.temaService.findById(postagem.tema.id)
+
+            if(!tema){
+                throw new HttpException("Tema não encontrado!", HttpStatus.NOT_FOUND)
+            }
+
+            return await this.postagemRepository.save(postagem)
+        }
+
         return await this.postagemRepository.save(postagem)
     }
 
     async updatePost(postagem: Postagem): Promise<Postagem>{
         
-        await this.findById(postagem.id)
+        let buscaPostagem: Postagem = await this.findById(postagem.id)
+
+        if(!buscaPostagem || !postagem.id){
+            throw new HttpException("Postagem não encontrada!", HttpStatus.NOT_FOUND)
+        }
+
+        if(postagem.tema){
+    
+            let tema = await this.temaService.findById(postagem.tema.id)
+
+            if(!tema){
+                throw new HttpException("Tema não encontrado!", HttpStatus.NOT_FOUND) 
+            }
+
+            return await this.postagemRepository.save(postagem)
+        }
 
         return await this.postagemRepository.save(postagem)
     }
     
     async delelePost(id: number): Promise<DeleteResult>{
 
-        await this.findById(id)
+        let buscaPostagem: Postagem = await this.findById(id)
+
+        if(!buscaPostagem){
+            throw new HttpException("Postagem não encontrada!", HttpStatus.NOT_FOUND) 
+        }
 
         return await this.postagemRepository.delete(id)
     }
